mosaic/mosaic-api/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-groq
pydantic
python-multipart
httpx
python-dotenv
requests
